updating conf
