+++ examples/qwen3/qat.yml
@@ -6,9 +6,22 @@ load_in_8bit: false
 load_in_4bit: false 
 strict: false
 
+plugins:
+  - axolotl.integrations.liger.LigerPlugin
+
+liger_glu_activation: true
+liger_rms_norm: true
+liger_layer_norm: true
+
+chat_template: qwen3
 datasets:
-  - path: tatsu-lab/alpaca
-    type: alpaca
+  - path: mlabonne/FineTome-100k
+    type: chat_template
+    split: train[:20%]
+    field_messages: conversations
+    message_property_mappings:
+      role: from
+      content: value
 dataset_prepared_path:
 val_set_size: 0.05
 output_dir: ./outputs/out
@@ -22,8 +35,8 @@ eval_sample_packing: false
 qat:
   activation_dtype: int8
   weight_dtype: int4
-  group_size: 32
-  fake_quant_after_n_steps: 100
+  group_size: 256
+  fake_quant_after_n_steps: 500
 
 wandb_project: qat
 wandb_entity:
@@ -31,7 +44,7 @@ wandb_watch:
 wandb_name:
 wandb_log_model:
 
-gradient_accumulation_steps: 4
+gradient_accumulation_steps: 1
 micro_batch_size: 4
 num_epochs: 1
 optimizer: adamw_torch_fused
