+++ src/axolotl/core/trainer_builder.py
@@ -59,6 +59,7 @@ from axolotl.core.training_args import (
 )
 from axolotl.integrations.base import PluginManager
 from axolotl.monkeypatch.multipack import SUPPORTED_MULTIPACK_MODEL_TYPES
+
 from axolotl.monkeypatch.relora import ReLoRACallback
 from axolotl.monkeypatch.trainer.lr import patch_trainer_get_lr
 from axolotl.processing_strategies import get_processing_strategy
@@ -76,6 +77,7 @@ from axolotl.utils.callbacks import (
 )
 from axolotl.utils.callbacks.lisa import lisa_callback_factory
 from axolotl.utils.callbacks.profiler import PytorchProfilerCallback
+from axolotl.utils.callbacks.qat import QATCallback
 from axolotl.utils.chat_templates import get_chat_template_from_config
 from axolotl.utils.collators import (
     BatchSamplerDataCollatorForSeq2Seq,
@@ -241,8 +243,8 @@ class HFCausalTrainerBuilder(TrainerBuilderBase):
             callbacks.append(ReLoRACallback(self.cfg))
 
         if (
-            hasattr(self.model, "use_bettertransformer")
-            and self.model.use_bettertransformer is True
+            hasattr(self.model, "use_bettertransformer") and
+            self.model.use_bettertransformer is True
         ):
             callbacks.append(SaveBetterTransformerModelCallback())
 
@@ -251,7 +253,9 @@ class HFCausalTrainerBuilder(TrainerBuilderBase):
 
         if self.cfg.gc_steps:
             callbacks.append(GCCallback(gc_steps=self.cfg.gc_steps))
-        
+
+        if self.cfg.qat and self.cfg.qat.fake_quant_after_n_steps:
+            callbacks.append(QATCallback(self.cfg.qat.fake_quant_after_n_steps))
         return callbacks
 
     def get_post_trainer_create_callbacks(self, trainer):
@@ -262,9 +266,9 @@ class HFCausalTrainerBuilder(TrainerBuilderBase):
             )
             callbacks.append(LogPredictionCallback(self.cfg))
         if (
-            self.cfg.use_mlflow
-            and is_mlflow_available()
-            and self.cfg.eval_table_size > 0
+            self.cfg.use_mlflow and
+            is_mlflow_available() and
+            self.cfg.eval_table_size > 0
         ):
             LogPredictionCallback = log_prediction_callback_factory(
                 trainer, self.tokenizer, "mlflow"
@@ -516,16 +520,16 @@ class HFCausalTrainerBuilder(TrainerBuilderBase):
         )
         training_arguments_kwargs["load_best_model_at_end"] = (
             (
-                self.cfg.load_best_model_at_end is not False
-                or self.cfg.early_stopping_patience
-            )
-            and (
-                (not self.cfg.test_datasets and self.cfg.val_set_size > 0)
-                or (self.cfg.test_datasets and self.cfg.val_set_size == 0)
-            )
-            and self.cfg.save_steps
-            and self.cfg.eval_steps
-            and self.cfg.save_steps % self.cfg.eval_steps == 0
+                self.cfg.load_best_model_at_end is not False or
+                self.cfg.early_stopping_patience
+            ) and
+            (
+                (not self.cfg.test_datasets and self.cfg.val_set_size > 0) or
+                (self.cfg.test_datasets and self.cfg.val_set_size == 0)
+            ) and
+            self.cfg.save_steps and
+            self.cfg.eval_steps and
+            self.cfg.save_steps % self.cfg.eval_steps == 0
         ) or False
 
         # handle ddp
@@ -834,8 +838,8 @@ class HFCausalTrainerBuilder(TrainerBuilderBase):
         else:
             trainer_kwargs["tokenizer"] = self.tokenizer
         if (
-            not (trainer_cls in [AxolotlRewardTrainer, AxolotlPRMTrainer])
-            and self.cfg.datasets is not None
+            not (trainer_cls in [AxolotlRewardTrainer, AxolotlPRMTrainer]) and
+            self.cfg.datasets is not None
         ):
             trainer_kwargs["dataset_tags"] = [
                 d["path"] for d in self.cfg.datasets if not Path(d["path"]).is_dir()
@@ -865,8 +869,8 @@ class HFCausalTrainerBuilder(TrainerBuilderBase):
     ):
         if training_args.pretraining:
             if (
-                self.cfg.pretraining_sample_concatenation is False
-                or self.cfg.micro_batch_size > 1
+                self.cfg.pretraining_sample_concatenation is False or
+                self.cfg.micro_batch_size > 1
             ):
                 return DataCollatorForSeq2Seq(self.tokenizer, **kwargs)
             return None
@@ -900,8 +904,8 @@ class HFCausalTrainerBuilder(TrainerBuilderBase):
             elif self.cfg.model_config_type in SUPPORTED_MULTIPACK_MODEL_TYPES:
                 collator = V2BatchSamplerDataCollatorForSeq2Seq
             elif (
-                self.cfg.model_config_type in ["llama"]
-                and self.cfg.flash_attention is not True
+                self.cfg.model_config_type in ["llama"] and
+                self.cfg.flash_attention is not True
             ):
                 collator = V2BatchSamplerDataCollatorForSeq2Seq
             else:
